implement BarCode component
