+++ form/src/main/java/org/dhis2/form/ui/provider/inputfield/FieldProvider.kt
@@ -46,8 +46,6 @@ import org.hisp.dhis.mobile.ui.designsystem.component.InputPercentage
 import org.hisp.dhis.mobile.ui.designsystem.component.InputPhoneNumber
 import org.hisp.dhis.mobile.ui.designsystem.component.InputPositiveInteger
 import org.hisp.dhis.mobile.ui.designsystem.component.InputPositiveIntegerOrZero
-import org.hisp.dhis.mobile.ui.designsystem.component.InputQRCode
-import org.hisp.dhis.mobile.ui.designsystem.component.InputText
 import org.hisp.dhis.mobile.ui.designsystem.component.internal.RegExValidations
 
 @OptIn(ExperimentalFoundationApi::class)
@@ -79,7 +77,7 @@ internal fun FieldProvider(
     if (fieldUiModel.optionSet == null) {
         when (fieldUiModel.valueType) {
             ValueType.TEXT -> {
-                ProvideFieldsForvalueTypeText(
+                ProvideInputsForValueTypeText(
                     modifier = modifierWithFocus,
                     fieldUiModel = fieldUiModel,
                     intentHandler = intentHandler,
@@ -303,48 +301,6 @@ internal fun FieldProvider(
                     },
                 )
             }
-        }
-    }
-}
-
-@Composable
-private fun ProvideFieldsForvalueTypeText(
-    modifier: Modifier = Modifier,
-    fieldUiModel: FieldUiModel,
-    intentHandler: (FormIntent) -> Unit,
-    uiEventHandler: (RecyclerViewUiEvents) -> Unit,
-) {
-    when (fieldUiModel.renderingType) {
-        UiRenderType.QR_CODE -> {
-            ProvideQRInput(
-                modifier = modifier,
-                fieldUiModel = fieldUiModel,
-                intentHandler = intentHandler,
-                uiEventHandler = uiEventHandler,
-            )
-        } else -> {
-            var value by remember {
-                mutableStateOf(fieldUiModel.value)
-            }
-            InputText(
-                modifier = modifier.fillMaxWidth(),
-                title = fieldUiModel.label,
-                state = fieldUiModel.inputState(),
-                supportingText = fieldUiModel.supportingText(),
-                legendData = fieldUiModel.legend(),
-                inputText = value ?: "",
-                isRequiredField = fieldUiModel.mandatory,
-                onValueChanged = {
-                    value = it
-                    intentHandler(
-                        FormIntent.OnTextChange(
-                            fieldUiModel.uid,
-                            value,
-                            fieldUiModel.valueType,
-                        ),
-                    )
-                },
-            )
         }
     }
 }
@@ -719,59 +675,6 @@ private fun ProvideInputLink(
                     fieldUiModel.uid,
                 ),
             )
-        },
-    )
-}
-
-@Composable
-private fun ProvideQRInput(
-    modifier: Modifier,
-    fieldUiModel: FieldUiModel,
-    intentHandler: (FormIntent) -> Unit,
-    uiEventHandler: (RecyclerViewUiEvents) -> Unit,
-) {
-    var value by remember(fieldUiModel.value) {
-        mutableStateOf(fieldUiModel.value)
-    }
-
-    InputQRCode(
-        modifier = modifier.fillMaxWidth(),
-        title = fieldUiModel.label,
-        state = fieldUiModel.inputState(),
-        supportingText = fieldUiModel.supportingText(),
-        legendData = fieldUiModel.legend(),
-        inputText = value ?: "",
-        isRequiredField = fieldUiModel.mandatory,
-        onValueChanged = {
-            value = it
-            intentHandler(
-                FormIntent.OnTextChange(
-                    fieldUiModel.uid,
-                    value,
-                    fieldUiModel.valueType,
-                ),
-            )
-        },
-        onQRButtonClicked = {
-            if (value.isNullOrEmpty()) {
-                uiEventHandler.invoke(
-                    RecyclerViewUiEvents.ScanQRCode(
-                        fieldUiModel.uid,
-                        optionSet = fieldUiModel.optionSet,
-                        fieldUiModel.renderingType,
-                    ),
-                )
-            } else {
-                uiEventHandler.invoke(
-                    RecyclerViewUiEvents.DisplayQRCode(
-                        fieldUiModel.uid,
-                        optionSet = fieldUiModel.optionSet,
-                        value = value!!,
-                        renderingType = fieldUiModel.renderingType,
-                        editable = fieldUiModel.editable,
-                    ),
-                )
-            }
         },
     )
 }
